create classes for the Configuration and the View of the loader
