+++ src/SwiftLoader/SwiftLoader.swift
@@ -24,7 +24,7 @@ public class SwiftLoader: UIView {
     private var title: String?
     private var speed = 1
     
-    private var config : Config = Config() {
+    private var config = SwiftLoaderConfig() {
         didSet {
             self.loadingView?.config = config
         }
@@ -38,7 +38,7 @@ public class SwiftLoader: UIView {
     
     class var sharedInstance: SwiftLoader {
         struct Singleton {
-            static let instance = SwiftLoader(frame: CGRectMake(0,0,Config().size,Config().size))
+            static let instance = SwiftLoader(frame: CGRectMake(0, 0, SwiftLoaderConfig().size, SwiftLoaderConfig().size))
         }
         return Singleton.instance
     }
@@ -84,7 +84,7 @@ public class SwiftLoader: UIView {
         loader.stop()
     }
     
-    public class func setConfig(config : Config) {
+    public class func setConfig(config : SwiftLoaderConfig) {
         let loader = SwiftLoader.sharedInstance
         loader.config = config
         loader.frame = CGRectMake(0, 0, loader.config.size, loader.config.size)
@@ -169,173 +169,7 @@ public class SwiftLoader: UIView {
         }
         return CGRectMake(loaderSpinnerMarginSide, loaderSpinnerMarginTop, loadingViewSize, loadingViewSize)
     }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.setup()
-    }
-    
-    required public init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-    }
-    
-    /**
-    *  Loader View
-    */
-    class SwiftLoadingView : UIView {
-        
-        private var speed : Int?
-        private var lineWidth : Float?
-        private var lineTintColor : UIColor?
-        private var backgroundLayer : CAShapeLayer?
-        private var isSpinning : Bool?
-        
-        private var config : Config = Config() {
-            didSet {
-                self.update()
-            }
-        }
-        
-        override init(frame: CGRect) {
-            super.init(frame: frame)
-            self.setup()
-        }
-        
-        required init?(coder aDecoder: NSCoder) {
-            super.init(coder: aDecoder)
-        }
-        
-        /**
-        Setup loading view
-        */
-        
-        private func setup() {
-            self.backgroundColor = UIColor.clearColor()
-            self.lineWidth = fmaxf(Float(self.frame.size.width) * 0.025, 1)
-            
-            self.backgroundLayer = CAShapeLayer()
-            self.backgroundLayer?.strokeColor = self.config.spinnerColor.CGColor
-            self.backgroundLayer?.fillColor = self.backgroundColor?.CGColor
-            self.backgroundLayer?.lineCap = kCALineCapRound
-            self.backgroundLayer?.lineWidth = CGFloat(self.lineWidth!)
-            self.layer.addSublayer(self.backgroundLayer!)
-        }
-        
-        private func update() {
-            self.lineWidth = self.config.spinnerLineWidth
-            self.speed = self.config.speed
-            
-            self.backgroundLayer?.lineWidth = CGFloat(self.lineWidth!)
-            self.backgroundLayer?.strokeColor = self.config.spinnerColor.CGColor
-        }
-        
-        /**
-        Draw Circle
-        */
-        
-        override func drawRect(rect: CGRect) {
-            self.backgroundLayer?.frame = self.bounds
-        }
-        
-        private func drawBackgroundCircle(partial : Bool) {
-            let startAngle : CGFloat = CGFloat(M_PI) / CGFloat(2.0)
-            var endAngle : CGFloat = (2.0 * CGFloat(M_PI)) + startAngle
-            
-            let center : CGPoint = CGPointMake(self.bounds.size.width / 2, self.bounds.size.height / 2)
-            let radius : CGFloat = (CGFloat(self.bounds.size.width) - CGFloat(self.lineWidth!)) / CGFloat(2.0)
-            
-            let processBackgroundPath : UIBezierPath = UIBezierPath()
-            processBackgroundPath.lineWidth = CGFloat(self.lineWidth!)
-            
-            if (partial) {
-                endAngle = (1.8 * CGFloat(M_PI)) + startAngle
-            }
-            
-            processBackgroundPath.addArcWithCenter(center, radius: radius, startAngle: startAngle, endAngle: endAngle, clockwise: true)
-            self.backgroundLayer?.path = processBackgroundPath.CGPath;
-        }
-        
-        /**
-        Start and stop spinning
-        */
-        
-        private func start() {
-            self.isSpinning? = true
-            self.drawBackgroundCircle(true)
-            
-            let rotationAnimation : CABasicAnimation = CABasicAnimation(keyPath: "transform.rotation.z")
-            rotationAnimation.toValue = NSNumber(double: M_PI * 2.0)
-            rotationAnimation.duration = CFTimeInterval(speed!);
-            rotationAnimation.cumulative = true;
-            rotationAnimation.repeatCount = HUGE;
-            self.backgroundLayer?.addAnimation(rotationAnimation, forKey: "rotationAnimation")
-        }
-        
-        private func stop() {
-            self.drawBackgroundCircle(false)
-            
-            self.backgroundLayer?.removeAllAnimations()
-            self.isSpinning? = false
         }
     }
     
-    
-    /**
-    * Loader config
-    */
-    public struct Config {
-        
-        /**
-        *  Size of loader
-        */
-        public var size : CGFloat = 120.0
-        
-        /**
-        *  Color of spinner view
-        */
-        public var spinnerColor = UIColor.blackColor()
-        
-        /**
-         *  Spinner Line Width
-         */
-        public var spinnerLineWidth :Float = 1.0
-        
-        /**
-        *  Color of title text
-        */
-        public var titleTextColor = UIColor.blackColor()
-        
-         /**
-         *  Speed of the spinner
-         */
-        public var speed :Int = 1
-        
-        /**
-        *  Font for title text in loader
-        */
-        public var titleTextFont : UIFont = UIFont.boldSystemFontOfSize(16.0)
-        
-        /**
-        *  Background color for loader
-        */
-        public var backgroundColor = UIColor.whiteColor()
-        
-        /**
-        *  Foreground color
-        */
-        public var foregroundColor = UIColor.clearColor()
-        
-        /**
-        *  Foreground alpha CGFloat, between 0.0 and 1.0
-        */
-        public var foregroundAlpha:CGFloat = 0.0
-        
-        /**
-        *  Corner radius for loader
-        */
-        public var cornerRadius : CGFloat = 10.0
-        
-        public init() {}
-        
-    }
 }
